cleanup sawyer stream stuff
