+++ OpenLocoTool/ObjectManager.cs
@@ -15,30 +15,10 @@ namespace OpenLocoTool
 			}
 		}
 
-		public static List<T> Get<T>(ObjectType type)
-			where T : ILocoStruct => Objects[type].Select(a => a.Object).Cast<T>().ToList();
+		public static IEnumerable<T> Get<T>(ObjectType type)
+			where T : ILocoStruct => Objects[type].Select(a => a.Object).Cast<T>();
 
 		public static void Add<T>(T obj) where T : ILocoObject
 			=> Objects[ObjectAttributes.ObjectType(obj.Object)].Add(obj);
 	}
-
-	// unused for now
-	//public class ObjectManager
-	//{
-	//	readonly Dictionary<ObjectType, List<ILocoObject>> Objects = new();
-
-	//	public ObjectManager()
-	//	{
-	//		foreach (var v in Enum.GetValues(typeof(ObjectType)))
-	//		{
-	//			Objects.Add((ObjectType)v, new List<ILocoObject>());
-	//		}
-	//	}
-
-	//	public List<T> Get<T>(ObjectType type)
-	//		where T : ILocoStruct => Objects[type].Select(a => a.Object).Cast<T>().ToList();
-
-	//	public void Add<T>(T obj) where T : ILocoObject
-	//		=> Objects[obj.S5Header.ObjectType].Add(obj);
-	//}
 }
